Ugly: Give TrebleApp write accesss to wifi configuration file
As part of O/O-MR1 vendor support for hostapd, an in-app IHostapd
implementation has been done. This needs to write to
/data/misc/wifi/hostapd.conf
/data/misc/wifi is originally in P writable only by wifi/wifi.
There is no easy way to gain "wifi" group access (though can be done
through platform.xml) from an app.
At the moment, all services requiring access to /data/misc/wifi are
running user "wifi", so this change is okay for the moment.

Change-Id: Ie9f40d6b6adee3403fc3fe0acf1c7c153c1893bb